Plug-in framework added.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3063 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Main Modules/ContrAppt.cs
@@ -3193,9 +3193,9 @@ namespace OpenDental{
 						break;
 				}
 			}
-			else if(e.Button.Tag.GetType()==typeof(int)){
+			else if(e.Button.Tag.GetType()==typeof(long)){
 				Patient pat=Patients.GetPat(PatCur.PatNum);
-				ProgramL.Execute((int)e.Button.Tag,pat);
+				ProgramL.Execute((long)e.Button.Tag,pat);
 			}
 		}
