+++ OpenDental/Main Modules/ContrDocs.cs
@@ -1004,8 +1004,8 @@ namespace OpenDental{
 						break;
 				}
 			}
-			else if(e.Button.Tag.GetType()==typeof(int)){
-				ProgramL.Execute((int)e.Button.Tag,PatCur);
+			else if(e.Button.Tag.GetType()==typeof(long)) {
+				ProgramL.Execute((long)e.Button.Tag,PatCur);
 			}
 		}
